Module model::graphql::backend

Interfaces provided by a backend data source consumed by the GraphQL model.

The entrypoint to this system of traits is DataSource, which describes the interface by which the GraphQL API interacts with the backend data provider. This is, in particular, the glue between the GraphQL and SQL models, since the SQL model will implement DataSource and the GraphQL layer will interact with the SQL layer exclusively through this trait.

A number of supporting traits are defined here which can be accessed through the DataSource trait by means of its associated types.

Structs

PageRequest

A specification of a page to load in a paginated connection.

Traits

Connection

A Relay-style paginated connection to a collection of objects.

DataSource

A source of data which can be served by the GraphQL API.

Type Definitions

Cursor

An index into Many.

Many

A one-to-many or many-to-many relationship to another Resource.