Derive Macro model::graphql::prelude::Query

#[derive(Query)]
{
    // Attributes available to this derive:
    #[query]
}

Derive resolvers for top-level query fields in a GraphQL API.

This macro generates an async_graphql #[Object] impl block for a struct, which includes resolvers for a number of given class types. These classes act as the entrypoints to an ontology, the rest of which is implied by the transitive closure of the entrypoint classes’ fields.

The classes to include as entrypoints are given using the class attribute, as in #[query(class(entrypoint: ClassType))]. entrypoint is the name of the GraphQL field which will be generated on the query object. It must be a Rust identifier (but will be converted from snake case to camel case in the GraphQL schema). ClassType is the type of items in the collection returned by entrypoint. You may specify multiple entrypoints that result to the same class type as long as they have different names.

Each entrypoint defined this way will generate a GraphQL field which takes the plural class inputs (a where predicate and Relay paging inputs) and produces a paginated connection of items of the class type matching the where clause.

Examples

use model::graphql::{prelude::*, EmptyMutation, EmptySubscription};

#[derive(Class)]
struct MyClass {
    #[class(primary)]
    field: u64,
}

#[derive(Query)]
#[query(class(classes: MyClass))]
struct Query;

let schema = Schema::build(Query, EmptyMutation, EmptySubscription).finish();

// Query by field.
schema
    .execute("{
        classes(
            where: {
                is: { is: { lit: 0 } }
            }
        ) {
            edges { node { field } }
        }
    }")
    .await
    .into_result()
    .unwrap();

Struct attributes

Attribute	Description	Type	Required

class	Class which should included in the ontology	field	yes