Trait model::graphql::traits::DataSource

pub trait DataSource {
    type Cursor<T: OutputType, E: ObjectType>: Cursor;
    type Connection<T: OutputType, C: ObjectType, E: ObjectType>: Connection<C>;
    type Error: Error;

    fn load_page<'life0, 'life1, 'async_trait, T, C, E>(
        &'life0 self,
        conn: &'life1 mut Self::Connection<T, C, E>,
        page: PageRequest<Self::Cursor<T, E>>
    ) -> Pin<Box<dyn Future<Output = Result<Vec<Edge<Self::Cursor<T, E>, T, E>>, Self::Error>> + Send + 'async_trait>>
    where
        T: 'async_trait + OutputType,
        C: 'async_trait + ObjectType,
        E: 'async_trait + ObjectType,
        Self: 'async_trait,
        'life0: 'async_trait,
        'life1: 'async_trait;
}

A source of data which can be served by the GraphQL API.

Required Associated Types

type Cursor<T: OutputType, E: ObjectType>: Cursor

An index into a paginated collection of objects.

The objects in the collection are of type T. Each object in the collection also represents a relationship, or edge, between the object which owns the collection and the object in the collection. These edges may have additional fields of type E, beyond the fields specified by Relay.

type Connection<T: OutputType, C: ObjectType, E: ObjectType>: Connection<C>

A Relay-style paginated connection to a collection of objects.

THe objects in the collection are of type T. Each object in the collection also represents a relationship, or edge, between the object which owns the collection and the object in the collection. These edges may have additional fields of type E, beyond the fields specified by Relay. The connection itself may also have additional fields of type C, beyond the fields specified by Relay.

type Error: Error

Errors reported while attempting to load data.

Required Methods

fn load_page<'life0, 'life1, 'async_trait, T, C, E>(
&'life0 self,
conn: &'life1 mut Self::Connection<T, C, E>,
page: PageRequest<Self::Cursor<T, E>>
) -> Pin<Box<dyn Future<Output = Result<Vec<Edge<Self::Cursor<T, E>, T, E>>, Self::Error>> + Send + 'async_trait>>where
T: 'async_trait + OutputType,
C: 'async_trait + ObjectType,
E: 'async_trait + ObjectType,
Self: 'async_trait,
'life0: 'async_trait,
'life1: 'async_trait,

Load a page from a paginated connection.