Module model::graphql::type_system

Type system for a relational GraphQL API.

A relational GraphQL API is one which allows the client to filter results based on complex queries that involve relations between nodes of the subgraph being queried. The queries are solved by compiling the expressive GraphQL query into a backend which can efficiently solve relational queries, like an RDBMS.

This module defines the types of objects that can appear in such an API along with a backend-agnostic interface for compiling them into low-level queries. Any API which follows this type system can be implemented by any backend which implements the relevant compilation traits. This is useful for testing (substituting a mock backend for the production one) and for avoiding lock-in, because it makes it easy to switch an existing application from one backend to another.

At a high-level the type system consists of the following notions:

Every entity in the application’s data model has a Type, including scalars (the leaves of the graph).

Every Type has a Predicate, a filter that can be applied to objects of that type in order to prune query output.

A Scalar is a subtype of Type which corresponds to a GraphQL scalar. These are the primitive types, the leaves of a graph. There is a fixed set of supported scalar types provided by this module, and these cannot be extended.

A Resource is a subtype of Type which corresponds to a GraphQL object type or a table in a relational database. Unlike a Scalar, whose only observable property is its value, a Resource can have many named properties of Scalar type as well as relations to other Resource types. These relations can be singular (one-to-one or many-to-one) or plural (one-to-many or many-to-many). They can be used in queries to easily select objects which related to a selected resource and to prune results based on a property of a relationship between resource.

Re-exports

pub use crate::typenum;

pub use crate::Array;

pub use crate::Length;

pub use resource::*;

pub use scalar::*;

Modules

resource

Resources are complex types in a relational GraphQL API.

scalar

In GraphQL, primitives are called scalars.

Macros

array

Create an Array from a list of items.

Traits

BuildError

An error encountered while reconstructing a GraphQL Type from query results.

Builder

A backend specific interface to query results, used to reconstruct a Type.

PluralPredicate

A boolean predicate on a PluralType consisting of items of type T.

PluralPredicateCompiler

A generic interface to a backend-specific plural predicate compiler.

PluralType

The type of a collection of items of a given Type.

Predicate

A boolean predicate on a Type T.

Type

The base type of the whole GraphQL type system.

Visitor

Visitor which allows a Type to describe itself to a backend.