Module model::graphql::type_system::resource

Resources are complex types in a relational GraphQL API.

A resource type is akin to an object type in GraphQL or a table in a relational database. It has its own fields as well as singular or plural relationships to other resources. Collections of items of a particular resource can be filter down using a ResourcePredicate. Entire collections of items of a given resource type can also be filtered in or out using a PluralPredicate.

Users can define their own resources by implementing the Resource trait and relatives, or by using the Resource derive macro.

Traits

Field

Metadata about a field of a resource.

FieldVisitor

Visitor which allows fields of a Resource to describe themselves to a backend.

PluralField

Metadata about a plural field of a resource.

PluralFieldVisitor

Visitor which allows plural fields of a Resource to describe themselves to a backend.

Resource

A complex type in the relational GraphQL type system.

ResourceBuilder

A backend specific interface to query results, used to reconstruct a Resource.

ResourcePredicate

A boolean predicate on a resource type T.

ResourceVisitor

Visitor which allows a Resource to describe itself to a backend.

Type Definitions

PluralFieldPredicate

The PluralPredicate used to filter a resource by its PluralField F.

Derive Macros

Resource

Derive an implementation of Resource, and related items, for a struct.