Trait model::graphql::type_system::resource::Field

pub trait Field: 'static {
    type Type: Type;
    type Resource: Resource;

    const NAME: &'static str;

    fn get(resource: &Self::Resource) -> &Self::Type;
    fn get_predicate(
        predicate: &<Self::Resource as Resource>::ResourcePredicate
    ) -> Option<&<Self::Type as Type>::Predicate>;
    fn get_predicate_mut(
        predicate: &mut <Self::Resource as Resource>::ResourcePredicate
    ) -> Option<&mut <Self::Type as Type>::Predicate>;
    fn take_predicate(
        predicate: &mut <Self::Resource as Resource>::ResourcePredicate
    ) -> Option<<Self::Type as Type>::Predicate>;

    fn is_id() -> bool { ... }
}

Metadata about a field of a resource.

Required Associated Types

type Type: Type

The type of the field.

type Resource: Resource

The resource that this field belongs to.

Required Associated Constants

const NAME: &'static str

The name of the field.

Required Methods

fn get(resource: &Self::Resource) -> &Self::Type

Access this field of a particular Resource.

fn get_predicate(
predicate: &<Self::Resource as Resource>::ResourcePredicate
) -> Option<&<Self::Type as Type>::Predicate>

Access the sub-predicate used to filter this field in a ResourcePredicate.

fn get_predicate_mut(
predicate: &mut <Self::Resource as Resource>::ResourcePredicate
) -> Option<&mut <Self::Type as Type>::Predicate>

Access the sub-predicate used to filter this field in a ResourcePredicate.

fn take_predicate(
predicate: &mut <Self::Resource as Resource>::ResourcePredicate
) -> Option<<Self::Type as Type>::Predicate>

Take the sub-predicate used to fitler this field in a ResourcePredicate.

The next time this function, get_predicate, or get_predicate_mut is called on the same field, it will return None, since the field has been taken.

Provided Methods

fn is_id() -> bool

Is this the unique ID field for its Resource?