Trait model::graphql::type_system::resource::PluralFieldVisitor

pub trait PluralFieldVisitor<T: Resource> {
    type Output;

    fn visit<F: PluralField<Resource = T>>(&mut self) -> Self::Output;
}

Visitor which allows plural fields of a Resource to describe themselves to a backend.

Required Associated Types

type Output

An output summarizing the results of visiting Ts plural fields.

Required Methods

fn visit<F: PluralField<Resource = T>>(&mut self) -> Self::Output

Tell the visitor about a field F.