Trait model::graphql::type_system::resource::Resource

pub trait Resource: Type<Predicate = Self::ResourcePredicate> {
    type ResourcePredicate: ResourcePredicate<Self>;

    fn build_resource<B: ResourceBuilder<Self>>(
        builder: B
    ) -> Result<Self, B::Error>;
}

A complex type in the relational GraphQL type system.

Required Associated Types

type ResourcePredicate: ResourcePredicate<Self>

Boolean predicates on this resource type.

This is always the same type as Predicate, but the alias ResourcePredicate has the more expressive trait bound ResourcePredicate instead of the generic Predicate.

Required Methods

fn build_resource<B: ResourceBuilder<Self>>(
builder: B
) -> Result<Self, B::Error>

Build a resource using a builder supplied by the backend.

This performs the same operation as build, but it can be called directly with a ResourceBuilder, instead of the more generic Builder. This is useful when it is known that a Type is actually a Resource.

It is an invariant that for all T: Resource, T::build(builder) == T::builder_resource(builder.resource()).