Trait model::graphql::type_system::resource::ResourcePredicate

pub trait ResourcePredicate<T: Resource>: Predicate<T> {
    fn compile_resource_predicate<C: ResourcePredicateCompiler<T>>(
        self,
        compiler: C
    ) -> C::Result;
}

A boolean predicate on a resource type T.

Required Methods

fn compile_resource_predicate<C: ResourcePredicateCompiler<T>>(
self,
compiler: C
) -> C::Result

Compile this predicate into a form which the backend can execute.

This performs the same operation as compile, but it can be called directly with a ResourcePredicateCompiler, instead of the more generic PredicateCompiler. This is useful when it is known that a Predicate is actually a ResourcePredicate.

It is an invariant that for all T: ResourcePredicate and x: T, x.compile(compiler) == x.compile_resource_predicate(compiler.resource()).