Trait model::graphql::type_system::resource::ResourceVisitor

pub trait ResourceVisitor<T: Resource>: Sized {
    type Output;

    fn visit_field_in_place<F: Field<Resource = T>>(&mut self);
    fn visit_plural_field_in_place<F: PluralField<Resource = T>>(&mut self);
    fn end(self) -> Self::Output;

    fn visit_field<F: Field<Resource = T>>(self) -> Self { ... }
    fn visit_plural_field<F: PluralField<Resource = T>>(self) -> Self { ... }
}

Visitor which allows a Resource to describe itself to a backend.

The implementation of this visitor will be backend-specific, but a Resource can use this backend-agnostic interface to describe its structure and fields to any backend.

Required Associated Types

type Output

An output summarizing the results of visiting T.

Required Methods

fn visit_field_in_place<F: Field<Resource = T>>(&mut self)

Tell the visitor about a field F of type T, mutating the visitor.

fn visit_plural_field_in_place<F: PluralField<Resource = T>>(&mut self)

Tell the visitor about a plural field F of type T, mutating the visitor.

fn end(self) -> Self::Output

Finish visiting the type and collect the output.

Provided Methods

fn visit_field<F: Field<Resource = T>>(self) -> Self

Tell the visitor about a field F of type T.

fn visit_plural_field<F: PluralField<Resource = T>>(self) -> Self

Tell the visitor about a plural field F of type T.