Trait model::graphql::type_system::Builder

pub trait Builder<T: Type> {
    type Error: BuildError;
    type Resource: ResourceBuilder<T, Error = Self::Error>
    where
        T: Resource;

    fn resource(self) -> Self::Resource
    where
        T: Resource;
    fn scalar(self) -> Result<T, Self::Error>
    where
        T: Scalar;
}

A backend specific interface to query results, used to reconstruct a Type.

Required Associated Types

type Error: BuildError

An error encountered while attempting to reconstruct the object.

type Resource: ResourceBuilder<T, Error = Self::Error>
where
T: Resource

A builder specifically for Resource types.

Required Methods

fn resource(self) -> Self::Resourcewhere
T: Resource,

Build a Resource.

fn scalar(self) -> Result<T, Self::Error>where
T: Scalar,

Reconstruct a Scalar.