Trait model::graphql::type_system::Type

pub trait Type: Clone + OutputType {
    type Predicate: Predicate<Self>;
    type PluralPredicate: PluralPredicate<Self>;
    type Input: Scalar;

    const NAME: &'static str;
    const PLURAL_NAME: &'static str;

    fn build<B: Builder<Self>>(builder: B) -> Result<Self, B::Error>;
    fn describe<V: Visitor<Self>>(visitor: V) -> V::Output;
}

The base type of the whole GraphQL type system.

All objects in a relational GraphQL API, from scalars to copmlext object types, have a type which is a subtype of Type, and this trait describes the functionality which is common to every entity in an API.

Required Associated Types

type Predicate: Predicate<Self>

A boolean predciate on objects of this type.

type PluralPredicate: PluralPredicate<Self>

A predicate on collections of objects of this type.

type Input: Scalar

The representation of values of this type when creating a new object.

For scalars, the input is just the value of the scalar itself. For resources, the input is the ID of the referenced resource.

Required Associated Constants

const NAME: &'static str

The name of this type.

This will usually be the stringified name of the type implementing this trait.

const PLURAL_NAME: &'static str

The plural form of Self::NAME.

Required Methods

fn build<B: Builder<Self>>(builder: B) -> Result<Self, B::Error>

Build an object of this type using a builder supplied by the backend.

This is used to reconstruct an object from a backend-specific query result.

fn describe<V: Visitor<Self>>(visitor: V) -> V::Output

Describe the structure and definition of this Type.