Module model::graphql::types

Types and traits describing the GraphQL protocol.

These types constitute the Rust bindings to the generic GraphQL protocol. They form the foundation upon which our specific GraphQL API is built.

Many of the types in this module are reexports from the [async_graphql] crate.

Re-exports

pub extern crate async_graphql;

Modules

connection

Types for Relay-compliant server

Structs

EmptyFields

Placeholder for connection objects (connections or edges) which have no additional fields.

EmptyMutation

Empty mutation

EmptySubscription

Empty subscription

Schema

GraphQL schema.

Traits

Class

A class type in the bill tracker GraphQL API.

InputType

Represents a GraphQL input type.

ObjectType

A GraphQL object.

OutputType

Represents a GraphQL output type.

Plural

A homogenous collection of items of Class type.

Scalar

A Scalar is a special type of Class which is represented as a GraphQL scalar.

Type Definitions

Context

Context object for resolve field

Result

An alias for Result<T, Error>.

Attribute Macros

Object

Derive Macros

Class

Derive an implementation of Class, and related items, for a struct.

InputObject

OneofObject

Query

Derive resolvers for top-level query fields in a GraphQL API.

SimpleObject