Trait model::graphql::types::Class

pub trait Class: OutputType {
    type Plural: Plural<Singular = Self>;
    type Predicate: InputType + Clone + Debug;
    type PluralPredicate: InputType + Clone + Debug;
}

A class type in the bill tracker GraphQL API.

A class type is akin to an object type in GraphQL or a table in a relational database. It has its own fields as well as singular or plural relationships to other classes. Collections of items of a particular class type can be filter down using a Predicate. Entire collections of items of this class type can also be filtered in or out using a PluralPredicate.

Required Associated Types

type Plural: Plural<Singular = Self>

A collection of items of this class.

type Predicate: InputType + Clone + Debug

A predicate on items of this class.

type PluralPredicate: InputType + Clone + Debug

A predicate on whole collections of this class.