Trait model::legiscan::Legiscan

pub trait Legiscan {
    type Dataset: Dataset;
    type DatasetMetadata: DatasetMetadata;

    fn list_datasets<'life0, 'async_trait>(
        &'life0 self,
        state: Option<State>,
        year: Option<u16>
    ) -> Pin<Box<dyn Future<Output = Result<Vec<Self::DatasetMetadata>, Error>> + Send + 'async_trait>>
    where
        Self: 'async_trait,
        'life0: 'async_trait;
    fn load_dataset<'life0, 'life1, 'async_trait>(
        &'life0 self,
        dataset: &'life1 Self::DatasetMetadata
    ) -> Pin<Box<dyn Future<Output = Result<Self::Dataset, Error>> + Send + 'async_trait>>
    where
        Self: 'async_trait,
        'life0: 'async_trait,
        'life1: 'async_trait;
}

A Legiscan client.

Required Associated Types

type Dataset: Dataset

A dataset, containing bulk information about many people and bills.

type DatasetMetadata: DatasetMetadata

Succinct metadata about a dataset, without the full contents.

Required Methods

fn list_datasets<'life0, 'async_trait>(
&'life0 self,
state: Option<State>,
year: Option<u16>
) -> Pin<Box<dyn Future<Output = Result<Vec<Self::DatasetMetadata>, Error>> + Send + 'async_trait>>where
Self: 'async_trait,
'life0: 'async_trait,

List available datasets, optionally filtering by state and year.

fn load_dataset<'life0, 'life1, 'async_trait>(
&'life0 self,
dataset: &'life1 Self::DatasetMetadata
) -> Pin<Box<dyn Future<Output = Result<Self::Dataset, Error>> + Send + 'async_trait>>where
Self: 'async_trait,
'life0: 'async_trait,
'life1: 'async_trait,

Load a dataset, including all bills, people, and votes involved.