Module model::schema

The schema describing the entities and relationships in the GraphQL API.

Modules

bill

district

issue

legislator

party

state

Structs

Bill

A piece of legislation.

District

A subdivision of a State with its own representatives in the state legislature.

Issue

A political issue.

Legislator

A state lawmaker.

Party

A political party.

Query

Entrypoint for read-only GraphQL queries.

State

A US state.

Functions

executor

Create an executor for the GraphQL API backed by a Postgres database.

generate

Create the schema for the GraphQL API.