Trait model::sql::db::CreateTable

pub trait CreateTable: Send {
    type Error: Error;

    fn constraint<I>(self, kind: ConstraintKind, columns: I) -> Self
    where
        I: IntoIterator,
        I::Item: Into<String>;
    fn execute<'async_trait>(
        self
    ) -> Pin<Box<dyn Future<Output = Result<(), Self::Error>> + Send + 'async_trait>>
    where
        Self: 'async_trait;
}

A CREATE TABLE statement which can be executed against the database.

Required Associated Types

type Error: Error

Errors returned by this statement.

Required Methods

fn constraint<I>(self, kind: ConstraintKind, columns: I) -> Selfwhere
I: IntoIterator,
I::Item: Into<String>,

Add a constraint to the table.

fn execute<'async_trait>(
self
) -> Pin<Box<dyn Future<Output = Result<(), Self::Error>> + Send + 'async_trait>>where
Self: 'async_trait,

Create the table.

This will execute a statement of the form CREATE TABLE IF NOT EXISTS table (columns constraints).

Errors

This method will fail if any of the specified constraints were invalid.s