Trait model::sql::db::Query

pub trait Query: Send {
    type Error: Error;
    type Row: Row<Error = Self::Error>;
    type Stream: Stream<Item = Result<Self::Row, Self::Error>> + Unpin + Send;

    fn clause(self, clause: Clause) -> Self;
    fn stream(self) -> Self::Stream;
}

A query which can be executed against the database.

Required Associated Types

type Error: Error

Errors returned by this query.

type Row: Row<Error = Self::Error>

Rows returned by this query.

type Stream: Stream<Item = Result<Self::Row, Self::Error>> + Unpin + Send

An asynchronous stream of rows.

Required Methods

fn clause(self, clause: Clause) -> Self

Add a clause to the query.

fn stream(self) -> Self::Stream

Run the query and get a stream of results.