Trait model::sql::db::SelectExt

pub trait SelectExt<'a>: Select<'a> {
    fn filter(
        self,
        column: impl Into<Column<'a>>,
        op: impl Into<Cow<'a, str>>,
        param: Value
    ) -> Self;
    fn join(
        self,
        table: impl Into<Cow<'a, str>>,
        lhs: impl Into<Column<'a>>,
        op: impl Into<Cow<'a, str>>,
        rhs: impl Into<Column<'a>>
    ) -> Self;
    fn clauses<I>(self, clauses: I) -> Self
    where
        I: IntoIterator,
        I::Item: Into<Clause<'a>>;
    fn one<'async_trait>(
        self
    ) -> Pin<Box<dyn Future<Output = Result<Self::Row, Self::Error>> + Send + 'async_trait>>
    where
        Self: 'async_trait;
    fn many<'async_trait>(
        self
    ) -> Pin<Box<dyn Future<Output = Result<Vec<Self::Row>, Self::Error>> + Send + 'async_trait>>
    where
        Self: 'async_trait;
    fn opt<'async_trait>(
        self
    ) -> Pin<Box<dyn Future<Output = Result<Option<Self::Row>, Self::Error>> + Send + 'async_trait>>
    where
        Self: 'async_trait;
}

An extension trait for Select that provides some higher-level functions.

Required Methods

fn filter(
self,
column: impl Into<Column<'a>>,
op: impl Into<Cow<'a, str>>,
param: Value
) -> Self

Add a WHERE clause to the query.

fn join(
self,
table: impl Into<Cow<'a, str>>,
lhs: impl Into<Column<'a>>,
op: impl Into<Cow<'a, str>>,
rhs: impl Into<Column<'a>>
) -> Self

Add a JOIN clause to the query.

fn clauses<I>(self, clauses: I) -> Selfwhere
I: IntoIterator,
I::Item: Into<Clause<'a>>,

Add multiple clauses to the query.

fn one<'async_trait>(
self
) -> Pin<Box<dyn Future<Output = Result<Self::Row, Self::Error>> + Send + 'async_trait>>where
Self: 'async_trait,

Run a query which is expected to return a single row.

Errors

This method will fail if the query does not return exactly one row.

fn many<'async_trait>(
self
) -> Pin<Box<dyn Future<Output = Result<Vec<Self::Row>, Self::Error>> + Send + 'async_trait>>where
Self: 'async_trait,

Run a query and collect the results.

fn opt<'async_trait>(
self
) -> Pin<Box<dyn Future<Output = Result<Option<Self::Row>, Self::Error>> + Send + 'async_trait>>where
Self: 'async_trait,

Run a query which is expected to return either 0 or 1 rows.

Errors

This method will fail if the query does not return exactly 0 or 1 rows.