Trait model::graphql::type_system::resource::Resource

pub trait Resource: Type<Predicate = Self::ResourcePredicate> {
    type NumFields: Length;
    type NumPluralFields: Length;
    type ResourcePredicate: ResourcePredicate<Self>;

    fn build_resource<B: ResourceBuilder<Self>>(
        builder: B
    ) -> Result<Self, B::Error>;
    fn describe_fields<V: FieldVisitor<Self>>(
        visitor: &mut V
    ) -> Array<V::Output, Self::NumFields>;
    fn describe_plural_fields<V: PluralFieldVisitor<Self>>(
        visitor: &mut V
    ) -> Array<V::Output, Self::NumPluralFields>;

    fn get<F: Field<Resource = Self>>(&self) -> &F::Type { ... }
    fn describe_resource<V: ResourceVisitor<Self>>(visitor: V) -> V::Output { ... }
    fn field_names() -> Array<&'static str, Self::NumFields> { ... }
}

A complex type in the relational GraphQL type system.

Required Associated Types

type NumFields: Length

The number of singular fields this resource has.

type NumPluralFields: Length

The number of plural fields this resource has.

type ResourcePredicate: ResourcePredicate<Self>

Boolean predicates on this resource type.

This is always the same type as Predicate, but the alias ResourcePredicate has the more expressive trait bound ResourcePredicate instead of the generic Predicate.

Required Methods

fn build_resource<B: ResourceBuilder<Self>>(
builder: B
) -> Result<Self, B::Error>

Build a resource using a builder supplied by the backend.

This performs the same operation as build, but it can be called directly with a ResourceBuilder, instead of the more generic Builder. This is useful when it is known that a Type is actually a Resource.

It is an invariant that for all T: Resource, T::build(builder) == T::builder_resource(builder.resource()).

fn describe_fields<V: FieldVisitor<Self>>(
visitor: &mut V
) -> Array<V::Output, Self::NumFields>

Describe the fields of this resource.

fn describe_plural_fields<V: PluralFieldVisitor<Self>>(
visitor: &mut V
) -> Array<V::Output, Self::NumPluralFields>

Describe the plural fields of this resource.

Provided Methods

fn get<F: Field<Resource = Self>>(&self) -> &F::Type

Access the field F of this Resource.

fn describe_resource<V: ResourceVisitor<Self>>(visitor: V) -> V::Output

Describe the structure and definition of this Resource.

This performs the same operation as describe, but it can be called directly with a ResourceVisitor, instead of the more generic Visitor. This is useful when it is known that a Type is actually a Resource.

It is an invariant that for all T: Resource, T::describe(visitor) == T::describe_resource(visitor.resource).

fn field_names() -> Array<&'static str, Self::NumFields>

The names of this resource’s singular fields.