Trait model::legiscan::Person

pub trait Person: Send + 'static {
    // Required methods
    fn id(&self) -> String;
    fn hash(&self) -> String;
    fn state(&self) -> State;
    fn party(&self) -> Party;
    fn name(&self) -> Name;
    fn district(&self) -> String;
}

Information about a person.

Required Methods

fn id(&self) -> String

The unique identifier for this person in the Legiscan API.

fn hash(&self) -> String

The hash of the information about this person.

This can be used to quickly check if a person needs to be updated, by storing the latest hash for each person and comparing it against the hash retrieved from Legiscan.

fn state(&self) -> State

The state where this person is a legislator.

fn party(&self) -> Party

The political party of which this person is a member.

fn name(&self) -> Name

The person’s full name.

fn district(&self) -> String

The name of the district this person represents.